refactor: move goal scroll logic from pager to page
- Removed direct RecordProvider reference from GoalPager
- Tracked current page index in GoalCalendarPage
- Simplified BottomRightButton by using index-based goal reference
